+++ NEWS.md
@@ -199,3 +199,4 @@
 
 # Racmacs 1.2.1
 * Titer merge method changed so that sd limit is implemented before any merging to < values
+* Expose further optimization options as user controllable options